requirements.txt
nltk==3.4.5
pandas==0.24.2
plotly==4.7.1
dash_html_components==1.0.3
yfinance==0.1.54
newspaper3k==0.2.8
dash_core_components==1.2.0
dash_bootstrap_components==0.9.2
textblob==0.15.3
#numpy==1.16.2
requests==2.22.0
beautifulsoup4==4.9.1
#newspaper==0.1.0.7
numpy==1.18.4
PriceIndices==1.1.1
gunicorn==20.0.4
dash_table==4.7.0
dash==1.12.0
scikit_learn==0.23.2
plotly-express==0.4.1
